requirement.txt
google.generativeai 
SpeechRecognition
emoji
texttospeech_v1 from google-cloud
pywhatkit
pyautogui
spotipy
